feat: update and delete files uploaded to s3 when updating its album

+++ src/main/java/com/untilled/roadcapture/api/client/AwsS3UploadClient.java
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.io.InputStream;
+import java.net.URL;
 import java.util.List;
+import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
@@ -31,14 +33,20 @@ public class AwsS3UploadClient implements UploadClient {
     }
 
     @Override
+    public boolean doesFileExists(String fileName) {
+        return amazonS3.doesObjectExist(this.bucket, fileName);
+    }
+
+    @Override
     public void deleteFiles(List<String> fileNames) {
-        fileNames.stream().forEach(fileName -> {
-            amazonS3.deleteObject(new DeleteObjectRequest(this.bucket, fileName));
-        });
+        fileNames.stream().filter(fileName -> doesFileExists(fileName))
+                .forEach(fileName -> amazonS3.deleteObject(new DeleteObjectRequest(this.bucket, fileName)));
     }
 
     @Override
     public void delete(String fileName) {
-        amazonS3.deleteObject(new DeleteObjectRequest(this.bucket, fileName));
+        if(doesFileExists(fileName)) {
+            amazonS3.deleteObject(new DeleteObjectRequest(this.bucket, fileName));
+        }
     }
 }
